[Behat] Remove ConfigurableProductFormTrait
