--- src/Sylius/Behat/Page/Admin/Product/ConfigurableProduct/ConfigurableProductFormTrait.php
@@ -1,84 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Sylius Sp. z o.o.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\Behat\Page\Admin\Product\ConfigurableProduct;
-
-use Behat\Mink\Element\NodeElement;
-use Sylius\Behat\Service\DriverHelper;
-
-trait ConfigurableProductFormTrait
-{
-    public function getDefinedFormElements(): array
-    {
-        return [
-            'channel' => '[data-test-channel-code="%channel_code%"]',
-            'channel_tab' => '[data-test-channel-tab="%channelCode%"]',
-            'channels' => '[data-test-channels]',
-            'code' => '[data-test-code]',
-            'enabled' => '[data-test-enabled]',
-            'form' => '[data-live-name-value="sylius_admin:product:form"]',
-            'product_options_autocomplete' => '[data-test-product-options-autocomplete]',
-            'side_navigation_tab' => '[data-test-side-navigation-tab="%name%"]',
-        ];
-    }
-
-    public function specifyCode(string $code): void
-    {
-        $this->getElement('code')->setValue($code);
-    }
-
-    public function specifyField(string $field, string $value): void
-    {
-        $this->getElement(lcfirst($field))->setValue($value);
-    }
-
-    public function selectOption(string $optionName): void
-    {
-        $this->changeTab('details');
-        $productOptionsAutocomplete = $this->getElement('product_options_autocomplete');
-
-        $this->autocompleteHelper->selectByName($this->getDriver(), $productOptionsAutocomplete->getXpath(), $optionName);
-    }
-
-    public function hasTab(string $name): bool
-    {
-        return $this->hasElement('side_navigation_tab', ['%name%' => $name]);
-    }
-
-    protected function getElement(string $name, array $parameters = []): NodeElement
-    {
-        if (!isset($parameters['%locale%'])) {
-            $parameters['%locale%'] = 'en_US';
-        }
-
-        return parent::getElement($name, $parameters);
-    }
-
-    private function waitForFormUpdate(): void
-    {
-        $form = $this->getElement('form');
-        sleep(1); // we need to sleep, as sometimes the check below is executed faster than the form sets the busy attribute
-        $form->waitFor(1500, function () use ($form) {
-            return !$form->hasAttribute('busy');
-        });
-    }
-
-    private function changeChannelTab(string $channelCode): void
-    {
-        if (DriverHelper::isNotJavascript($this->getDriver())) {
-            return;
-        }
-
-        $this->getElement('channel_tab', ['%channelCode%' => $channelCode])->click();
-    }
-}
