+++ src/Sylius/Behat/Page/Admin/Product/ConfigurableProduct/CreateConfigurableProductPage.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Sylius\Behat\Page\Admin\Product\ConfigurableProduct;
 
+use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Session;
 use Sylius\Behat\Page\Admin\Crud\CreatePage as BaseCreatePage;
 use Sylius\Behat\Page\Admin\Product\Common\ProductAttributesTrait;
@@ -24,7 +25,6 @@ use Symfony\Component\Routing\RouterInterface;
 
 class CreateConfigurableProductPage extends BaseCreatePage implements CreateConfigurableProductPageInterface
 {
-    use ConfigurableProductFormTrait;
     use ProductAttributesTrait;
     use ProductMediaTrait;
     use ProductTaxonomyTrait;
@@ -47,6 +47,24 @@ class CreateConfigurableProductPage extends BaseCreatePage implements CreateConf
         parent::create();
     }
 
+    public function specifyCode(string $code): void
+    {
+        $this->getElement('code')->setValue($code);
+    }
+
+    public function specifyField(string $field, string $value): void
+    {
+        $this->getElement(lcfirst($field))->setValue($value);
+    }
+
+    public function selectOption(string $optionName): void
+    {
+        $this->changeTab('details');
+        $productOptionsAutocomplete = $this->getElement('product_options_autocomplete');
+
+        $this->autocompleteHelper->selectByName($this->getDriver(), $productOptionsAutocomplete->getXpath(), $optionName);
+    }
+
     /**
      * @return string[]
      */
@@ -54,11 +72,38 @@ class CreateConfigurableProductPage extends BaseCreatePage implements CreateConf
     {
         return array_merge(
             parent::getDefinedElements(),
-            $this->getDefinedFormElements(),
             $this->getDefinedProductMediaElements(),
             $this->getDefinedProductAttributesElements(),
             $this->getDefinedProductTranslationsElements(),
             $this->getDefinedProductTaxonomyElements(),
+            [
+                'channel' => '[data-test-channel-code="%channel_code%"]',
+                'channel_tab' => '[data-test-channel-tab="%channelCode%"]',
+                'channels' => '[data-test-channels]',
+                'code' => '[data-test-code]',
+                'enabled' => '[data-test-enabled]',
+                'form' => '[data-live-name-value="sylius_admin:product:form"]',
+                'product_options_autocomplete' => '[data-test-product-options-autocomplete]',
+                'side_navigation_tab' => '[data-test-side-navigation-tab="%name%"]',
+            ],
         );
+    }
+
+    protected function getElement(string $name, array $parameters = []): NodeElement
+    {
+        if (!isset($parameters['%locale%'])) {
+            $parameters['%locale%'] = 'en_US';
+        }
+
+        return parent::getElement($name, $parameters);
+    }
+
+    private function waitForFormUpdate(): void
+    {
+        $form = $this->getElement('form');
+        sleep(1); // we need to sleep, as sometimes the check below is executed faster than the form sets the busy attribute
+        $form->waitFor(1500, function () use ($form) {
+            return !$form->hasAttribute('busy');
+        });
     }
 }
